requirements.txt
Django==1.4.3
Pillow==1.7.8
django-disqus==0.4.1
django-filebrowser==3.5.1
django-flatpages-tinymce==0.1.1
django-grappelli==2.4.3
django-tinymce==1.5.1b4
wsgiref==0.1.2
